gpu: jit: conv: fix conversion warning

+++ src/gpu/jit/conv/config.cpp
@@ -2421,7 +2421,7 @@ std::string conv_config_t::str() const {
 
 std::string pad_str(std::string s, int pad) {
     auto pos = (pad >= 0 ? 0 : s.length());
-    int off = std::abs(pad) - s.length();
+    int off = std::abs(pad) - (int)s.length();
     s.insert(pos, std::max(off, 0), ' ');
     return s;
 }
